+++ module-admin/src/main/java/org/example/AdminApplication.java
@@ -4,14 +4,10 @@ import jakarta.annotation.PostConstruct;
 import java.util.TimeZone;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.scheduling.annotation.EnableScheduling;
 
 @EnableJpaAuditing
-@EnableScheduling
 @SpringBootApplication
-@EnableFeignClients
 public class AdminApplication {
 
   @PostConstruct
