+++ module-admin/src/main/java/org/example/domain/workbook_problem/repository/ListWorkbookProblemRepository.java
@@ -6,8 +6,6 @@ import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.example.domain.study.Study;
-import org.example.domain.week.Week;
 import org.example.domain.workbook_problem.controller.response.ListWorkbookProblemDto;
 import org.springframework.stereotype.Repository;
 
@@ -17,17 +15,6 @@ public class ListWorkbookProblemRepository {
 
   private final JPAQueryFactory queryFactory;
 
-
-  public List<Integer> getWorkbookProblemList(Week week, Study study) {
-    return queryFactory
-      .select(workbookProblem.problem.number)
-      .from(workbookProblem)
-      .where(
-        workbookProblem.workbook.week.eq(week),
-        workbookProblem.workbook.study.eq(study)
-      )
-      .fetch();
-  }
 
   public List<ListWorkbookProblemDto> getWorkbookProblemList(Long workbookId) {
     return queryFactory
