+++ module-user/src/main/java/org/example/domain/algorithm/service/AlgorithmService.java
