+++ module-user/src/main/java/org/example/domain/attendance/service/AttendanceService.java
@@ -8,7 +8,12 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class AttendanceService {
 
+  private final CreateAttendanceService createAttendanceService;
   private final ListAttendanceService listAttendanceService;
+
+  public void createAttendance() {
+    createAttendanceService.createAttendance();
+  }
 
   /**
    * 정규 스터디 출석부 조회
