+++ module-user/src/main/java/org/example/domain/attendance_request_problem/repository/ListAttendanceRequestProblemRepository.java
@@ -5,6 +5,7 @@ import static org.example.domain.attendance_request_problem.QAttendanceRequestPr
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.example.domain.attendance_request.AttendanceRequest;
 import org.springframework.stereotype.Repository;
 
 @Repository
@@ -13,6 +14,17 @@ public class ListAttendanceRequestProblemRepository {
 
   private final JPAQueryFactory queryFactory;
 
+  public int getRequestCount(AttendanceRequest attendanceRequest) {
+    return queryFactory
+      .select(attendanceRequestProblem)
+      .from(attendanceRequestProblem)
+      .where(
+        attendanceRequestProblem.attendanceRequest.eq(attendanceRequest),
+        attendanceRequestProblem.problemUrl.contains("acmicpc.net").not(), // 백준 제외
+        attendanceRequestProblem.problemUrl.contains("boj.kr").not() // 백준 제외
+      )
+      .fetch().size();
+  }
 
   public List<String> getAttendanceRequestProblemList(Long attendanceRequestId) {
     return queryFactory
