+++ module-user/src/main/java/org/example/domain/study/repository/ListStudyRepository.java
@@ -11,6 +11,7 @@ import com.querydsl.jpa.JPAExpressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.example.domain.study.controller.response.ListRegularStudyDto;
 import org.example.domain.study.controller.response.ListStudyDto;
 import org.example.domain.study.enums.StudyType;
 import org.example.domain.study_member.enums.StudyMemberRole;
@@ -23,6 +24,40 @@ import org.springframework.stereotype.Repository;
 public class ListStudyRepository {
 
   private final JPAQueryFactory queryFactory;
+
+  /**
+   * 정규 스터디 목록 조회
+   */
+  public List<ListRegularStudyDto> getRegularStudyList(boolean currentGeneration) {
+    Integer maxGeneration = queryFactory
+      .select(generation.value.max())
+      .from(generation)
+      .fetchOne();
+
+    return queryFactory
+      .select(Projections.fields(ListRegularStudyDto.class,
+          study.id.as("studyId"),
+          study.profileUrl,
+          study.name,
+          Expressions.as(
+            JPAExpressions
+              .select(studyMember.count())
+              .from(studyMember)
+              .where(
+                studyMember.study.eq(study),
+                studyMember.status.eq(StudyMemberStatus.PASS)
+              )
+            , "memberCount")
+        )
+      )
+      .from(study)
+      .where(
+        study.type.eq(StudyType.REGULAR),
+        study.generation.value.goe(currentGeneration ? maxGeneration : Integer.valueOf(16)) // 서비스 적용 기수 : 16기
+      )
+      .orderBy(study.generation.value.desc())
+      .fetch();
+  }
 
   /**
    * 스터디 목록 조회
