+++ module-user/src/main/java/org/example/domain/workbook_problem/repository/ListWorkbookProblemRepository.java
@@ -8,6 +8,8 @@ import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.example.domain.study.Study;
+import org.example.domain.week.Week;
 import org.example.domain.workbook_problem.controller.response.ListWorkbookProblemDto;
 import org.example.util.http_request.Url;
 import org.springframework.stereotype.Repository;
@@ -18,6 +20,16 @@ public class ListWorkbookProblemRepository {
 
   private final JPAQueryFactory queryFactory;
 
+  public List<Integer> getWorkbookProblemList(Week week, Study study) {
+    return queryFactory
+      .select(workbookProblem.problem.number)
+      .from(workbookProblem)
+      .where(
+        workbookProblem.workbook.week.eq(week),
+        workbookProblem.workbook.study.eq(study)
+      )
+      .fetch();
+  }
 
   public List<ListWorkbookProblemDto> getWorkbookProblemList(Long workbookId) {
     return queryFactory
