+++ module-user/src/main/java/org/example/schedule/solved_ac/response/algorithm/AlgorithmDto.java
