+++ module-user/src/main/resources/application.yml
@@ -1,8 +1,11 @@
 
-
 spring:
   config:
     import: application-core.yml
+
+feign:
+  solved-ac:
+    url: https://solved.ac/api/v3
 
 springdoc:
   swagger-ui:
